refactor: move storage implementations to their own files
